Fix checkstyle, move to `common.logging` / Address comments by @trustin
