+++ core/src/main/java/com/linecorp/armeria/common/logging/BuiltInProperties.java
@@ -13,21 +13,21 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.linecorp.armeria.common;
+package com.linecorp.armeria.common.logging;
 
-import static com.linecorp.armeria.common.BuiltInProperty.CLIENT_IP;
-import static com.linecorp.armeria.common.BuiltInProperty.LOCAL_HOST;
-import static com.linecorp.armeria.common.BuiltInProperty.LOCAL_IP;
-import static com.linecorp.armeria.common.BuiltInProperty.LOCAL_PORT;
-import static com.linecorp.armeria.common.BuiltInProperty.REMOTE_HOST;
-import static com.linecorp.armeria.common.BuiltInProperty.REMOTE_IP;
-import static com.linecorp.armeria.common.BuiltInProperty.REMOTE_PORT;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_RPC_METHOD;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_RPC_PARAMS;
-import static com.linecorp.armeria.common.BuiltInProperty.RES_RPC_RESULT;
-import static com.linecorp.armeria.common.BuiltInProperty.TLS_CIPHER;
-import static com.linecorp.armeria.common.BuiltInProperty.TLS_PROTO;
-import static com.linecorp.armeria.common.BuiltInProperty.TLS_SESSION_ID;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.CLIENT_IP;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.LOCAL_HOST;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.LOCAL_IP;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.LOCAL_PORT;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REMOTE_HOST;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REMOTE_IP;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REMOTE_PORT;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_RPC_METHOD;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_RPC_PARAMS;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.RES_RPC_RESULT;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.TLS_CIPHER;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.TLS_PROTO;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.TLS_SESSION_ID;
 
 final class BuiltInProperties {
