+++ core/src/main/java/com/linecorp/armeria/common/logging/RequestContextExporter.java
@@ -14,34 +14,34 @@
  * under the License.
  */
 
-package com.linecorp.armeria.common;
+package com.linecorp.armeria.common.logging;
 
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
-import static com.linecorp.armeria.common.BuiltInProperty.CLIENT_IP;
-import static com.linecorp.armeria.common.BuiltInProperty.ELAPSED_NANOS;
-import static com.linecorp.armeria.common.BuiltInProperty.LOCAL_HOST;
-import static com.linecorp.armeria.common.BuiltInProperty.LOCAL_IP;
-import static com.linecorp.armeria.common.BuiltInProperty.LOCAL_PORT;
-import static com.linecorp.armeria.common.BuiltInProperty.REMOTE_HOST;
-import static com.linecorp.armeria.common.BuiltInProperty.REMOTE_IP;
-import static com.linecorp.armeria.common.BuiltInProperty.REMOTE_PORT;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_AUTHORITY;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_CONTENT_LENGTH;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_DIRECTION;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_METHOD;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_PATH;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_QUERY;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_RPC_METHOD;
-import static com.linecorp.armeria.common.BuiltInProperty.REQ_RPC_PARAMS;
-import static com.linecorp.armeria.common.BuiltInProperty.RES_CONTENT_LENGTH;
-import static com.linecorp.armeria.common.BuiltInProperty.RES_RPC_RESULT;
-import static com.linecorp.armeria.common.BuiltInProperty.RES_STATUS_CODE;
-import static com.linecorp.armeria.common.BuiltInProperty.SCHEME;
-import static com.linecorp.armeria.common.BuiltInProperty.TLS_CIPHER;
-import static com.linecorp.armeria.common.BuiltInProperty.TLS_PROTO;
-import static com.linecorp.armeria.common.BuiltInProperty.TLS_SESSION_ID;
-import static com.linecorp.armeria.common.RequestContextExporterBuilder.PREFIX_ATTRS;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.CLIENT_IP;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.ELAPSED_NANOS;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.LOCAL_HOST;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.LOCAL_IP;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.LOCAL_PORT;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REMOTE_HOST;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REMOTE_IP;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REMOTE_PORT;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_AUTHORITY;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_CONTENT_LENGTH;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_DIRECTION;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_METHOD;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_PATH;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_QUERY;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_RPC_METHOD;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.REQ_RPC_PARAMS;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.RES_CONTENT_LENGTH;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.RES_RPC_RESULT;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.RES_STATUS_CODE;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.SCHEME;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.TLS_CIPHER;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.TLS_PROTO;
+import static com.linecorp.armeria.common.logging.BuiltInProperty.TLS_SESSION_ID;
+import static com.linecorp.armeria.common.logging.RequestContextExporterBuilder.PREFIX_ATTRS;
 import static java.util.Objects.requireNonNull;
 
 import java.net.InetAddress;
@@ -62,8 +62,12 @@ import com.google.common.io.BaseEncoding;
 
 import com.linecorp.armeria.client.ClientRequestContext;
 import com.linecorp.armeria.client.Endpoint;
-import com.linecorp.armeria.common.logging.RequestLog;
-import com.linecorp.armeria.common.logging.RequestLogAvailability;
+import com.linecorp.armeria.common.HttpHeaderNames;
+import com.linecorp.armeria.common.HttpHeaders;
+import com.linecorp.armeria.common.HttpRequest;
+import com.linecorp.armeria.common.RequestContext;
+import com.linecorp.armeria.common.RpcRequest;
+import com.linecorp.armeria.common.RpcResponse;
 import com.linecorp.armeria.server.ServiceRequestContext;
 
 import io.netty.util.AsciiString;
