+++ core/src/test/java/com/linecorp/armeria/common/logging/RequestContextExporterBuilderTest.java
@@ -14,7 +14,7 @@
  * under the License.
  */
 
-package com.linecorp.armeria.common;
+package com.linecorp.armeria.common.logging;
 
 import static org.assertj.core.api.Assertions.assertThat;
